docs(adr-063): record Phase 5 complete; split fact_codec under the size ceiling
Phase 5's field-by-field conversion is done, so ADR-063's Status paragraph,
its adr/index.md row and the plan's Phase 5 section say so — with the
per-batch detail kept in the ADR alone, so the two cannot drift.
KNOWN_UNCONVERTED_ELIGIBLE_FACTS' now-empty halves document why they stay
declared rather than being deleted with their last entry.

The tenth batch pushed storage/fact_codec.py past ADR-061's 800-line
production ceiling; the case-(a) legacy-load corrections move to
storage/fact_backfill.py and the per-field schema-version thresholds both
sides need to a storage/fact_schema_versions.py leaf, so no import cycle
forms. fact_codec re-exports the moved names unchanged.
