+++ abicheck/storage/fact_codec.py
@@ -25,14 +25,38 @@ concern out of `serialization.py`'s neighbours.
 
 from __future__ import annotations
 
-from collections.abc import Iterator
-from dataclasses import dataclass, replace
+from dataclasses import replace
 from typing import TYPE_CHECKING, Any
 
 from ..model import AccessLevel, Fact, FactStatus, SymbolBinding
 
+# ADR-061's 800-line production ceiling: the case-(a) legacy-load
+# corrections live in ``fact_backfill.py`` and are re-exported here, so
+# ``serialization.py`` and every test importing them from this module are
+# unaffected by that split. One-directional: ``fact_backfill`` imports the
+# per-field schema-version thresholds from the ``fact_schema_versions.py``
+# leaf, never from here, so this re-export forms no cycle.
+from .fact_backfill import (
+    CaseAFactRule as CaseAFactRule,
+    apply_case_a_fact_backfill as apply_case_a_fact_backfill,
+    apply_legacy_fact_backfill as apply_legacy_fact_backfill,
+)
+from .fact_schema_versions import (
+    _FACT_FIELDS_SCHEMA_VERSION,
+    _MIN_SCHEMA_VERSION_FOR_DEPRECATION_FACTS,
+    _MIN_SCHEMA_VERSION_FOR_ENUMTYPE_FACTS,
+    _MIN_SCHEMA_VERSION_FOR_FUNCTION_CASE_B_FACTS,
+    _MIN_SCHEMA_VERSION_FOR_IS_FINAL_FACT,
+    _MIN_SCHEMA_VERSION_FOR_LAST_CASE_A_FACTS,
+    _MIN_SCHEMA_VERSION_FOR_RECORDTYPE_CASE_B_FACTS,
+    _MIN_SCHEMA_VERSION_FOR_SNAPSHOT_CASE_B_FACTS,
+    _MIN_SCHEMA_VERSION_FOR_TYPEFIELD_CV_FACTS,
+    _MIN_SCHEMA_VERSION_FOR_TYPEFIELD_VALUE_FACTS,
+    _MIN_SCHEMA_VERSION_FOR_VARIABLE_CASE_B_FACTS,
+)
+
 if TYPE_CHECKING:
-    from ..model import Function, RecordType
+    pass
 
 __all__ = [
     "CaseAFactRule",
@@ -189,44 +213,6 @@ def _encode_one(fact_dict: dict[str, Any] | None) -> None:
     status = fact_dict.get("status")
     if isinstance(status, FactStatus):
         fact_dict["status"] = status.value
-
-
-# The schema_version this phase bumped SCHEMA_VERSION to when it started
-# persisting a *_fact sibling for every legacy field it emits (serialization.py).
-_FACT_FIELDS_SCHEMA_VERSION = 26
-
-# ADR-063 Phase 5: the schema_version RecordType.is_final_fact started being
-# persisted at — independent of _FACT_FIELDS_SCHEMA_VERSION above, since a
-# document between the two thresholds (v26..v29) genuinely never carried
-# this key at all, the same way a pre-v26 document never carried any *_fact
-# key. See decode_fact's own docstring for why this threshold matters.
-_MIN_SCHEMA_VERSION_FOR_IS_FINAL_FACT = 30
-
-# ADR-063 Phase 5 (second batch): the schema_version RecordType's remaining
-# case-(b) *_fact siblings (is_abstract/data_size_bits/is_standard_layout/
-# is_trivially_copyable/qualified_name/source_header) started being
-# persisted at — one shared threshold, since all six land together in the
-# same schema bump. Same reasoning as _MIN_SCHEMA_VERSION_FOR_IS_FINAL_FACT
-# above: a document below this version never carried these keys at all.
-_MIN_SCHEMA_VERSION_FOR_RECORDTYPE_CASE_B_FACTS = 32
-
-# ADR-063 Phase 5 (third batch): the schema_version EnumType's own
-# qualified_name_fact/source_header_fact siblings started being persisted
-# at.
-_MIN_SCHEMA_VERSION_FOR_ENUMTYPE_FACTS = 33
-
-# ADR-063 Phase 5 (fourth batch): the schema_version Variable's own
-# source_header_fact/alignment_bits_fact/elf_binding_fact siblings started
-# being persisted at.
-_MIN_SCHEMA_VERSION_FOR_VARIABLE_CASE_B_FACTS = 34
-
-# ADR-063 Phase 5 (fifth batch): the schema_version Function's own ten
-# case-(b) *_fact siblings started being persisted at.
-_MIN_SCHEMA_VERSION_FOR_FUNCTION_CASE_B_FACTS = 35
-
-# ADR-063 Phase 5 (sixth batch): the schema_version AbiSnapshot's own
-# ast_resolved_standard_fact sibling started being persisted at.
-_MIN_SCHEMA_VERSION_FOR_SNAPSHOT_CASE_B_FACTS = 36
 
 
 def decode_fact(
@@ -488,33 +474,6 @@ def decode_snapshot_facts(d: dict[str, Any], schema_version: int) -> dict[str, A
     }
 
 
-# ADR-063 Phase 5 (eighth batch): the schema_version TypeField's own
-# case-(a) is_const_fact/is_volatile_fact/is_mutable_fact siblings started
-# being persisted at.
-_MIN_SCHEMA_VERSION_FOR_TYPEFIELD_CV_FACTS = 38
-
-# The same threshold for TypeField's other two case-(a) fields (`default`,
-# `deprecated`) -- they land in the same schema bump, but each is guarded by
-# its own reliability flag (clang_field_initializer_facts_reliable /
-# clang_deprecation_facts_reliable), so they are named separately here
-# rather than folded into the CV constant above.
-_MIN_SCHEMA_VERSION_FOR_TYPEFIELD_VALUE_FACTS = 38
-
-# ADR-063 Phase 5 (ninth batch): the schema_version the `deprecated` family
-# (Function/Variable/RecordType/EnumType -- TypeField's own landed one batch
-# earlier, at v38) and EnumType.is_scoped started being persisted at. All
-# five are case (a), guarded by the one flag that already covers them:
-# AbiSnapshot.clang_deprecation_facts_reliable.
-_MIN_SCHEMA_VERSION_FOR_DEPRECATION_FACTS = 39
-
-# ADR-063 Phase 5 (tenth batch): the schema_version Param.is_restrict_fact
-# and Variable.access_fact started being persisted at -- the last two
-# entries of this phase's own KNOWN_UNCONVERTED_ELIGIBLE_FACTS allowlist,
-# guarded by clang_restrict_facts_reliable and
-# castxml_var_access_facts_reliable respectively.
-_MIN_SCHEMA_VERSION_FOR_LAST_CASE_A_FACTS = 40
-
-
 def decode_fact_with_legacy_presence(
     owner_dict: dict[str, Any],
     legacy_key: str,
@@ -606,295 +565,3 @@ def decode_field_facts(fld: dict[str, Any], schema_version: int) -> dict[str, An
             _MIN_SCHEMA_VERSION_FOR_TYPEFIELD_VALUE_FACTS,
         ),
     }
-
-
-@dataclass(frozen=True)
-class CaseAFactRule:
-    """One case-(a) field's legacy-load correction (ADR-063 Phase 5).
-
-    ``owner``/``field`` name the legacy field; ``min_schema_version`` is the
-    schema_version that field's own ``<field>_fact`` sibling started being
-    persisted at; ``reliable`` is this snapshot's already-resolved answer to
-    "is the flag guarding this field's availability trustworthy here"
-    (``serialization.py`` computes every ``*_facts_reliable`` value, folding
-    in any producer gate — see ``apply_legacy_fact_backfill``'s own
-    ``ast_producer`` note); ``normalized_default`` is the value the legacy
-    field is reset to when the fact is downgraded, so the pair cannot be
-    left holding a placeholder beside a NOT_COLLECTED status.
-    """
-
-    owner: str
-    field: str
-    min_schema_version: int
-    reliable: bool
-    normalized_default: Any
-
-
-def _owner_pairs(
-    d: dict[str, Any],
-    owner: str,
-    decoded: dict[str, list[Any]],
-) -> Iterator[tuple[dict[str, Any], Any]]:
-    """Every ``(raw dict, decoded object)`` pair for one *owner* dataclass.
-
-    The one place this module knows how a given owner's instances are
-    reached from the raw snapshot document — two owners (``TypeField``,
-    ``Param``) live one level below a collection rather than in one, and a
-    per-field ``zip`` open-coded at each call site is exactly the kind of
-    duplication a later owner's conversion would get subtly wrong.
-    """
-    if owner == "RecordType":
-        yield from zip(d.get("types", []), decoded.get("types", []), strict=False)
-    elif owner == "EnumType":
-        yield from zip(d.get("enums", []), decoded.get("enums", []), strict=False)
-    elif owner == "Variable":
-        yield from zip(
-            d.get("variables", []), decoded.get("variables", []), strict=False
-        )
-    elif owner == "Function":
-        yield from zip(
-            d.get("functions", []), decoded.get("functions", []), strict=False
-        )
-    elif owner == "TypeField":
-        for type_dict, record in zip(
-            d.get("types", []), decoded.get("types", []), strict=False
-        ):
-            yield from zip(type_dict.get("fields", []), record.fields, strict=False)
-    elif owner == "Param":
-        for func_dict, func in zip(
-            d.get("functions", []), decoded.get("functions", []), strict=False
-        ):
-            yield from zip(func_dict.get("params", []), func.params, strict=False)
-    else:  # pragma: no cover - guarded by the caller's own closed rule set
-        raise ValueError(f"no raw-document navigation known for owner {owner!r}")
-
-
-def apply_case_a_fact_backfill(
-    d: dict[str, Any],
-    *,
-    schema_version: int,
-    rules: tuple[CaseAFactRule, ...],
-    **decoded: list[Any],
-) -> None:
-    """Downgrade every case-(a) fact a legacy document cannot vouch for.
-
-    A document below a field's own ``min_schema_version`` carries no
-    ``<field>_fact`` key at all, so the owning dataclass's ``__post_init__``
-    bridge already backfilled ``Fact.present(raw_value)`` — correct when the
-    snapshot-level reliability flag guarding that field says this producer's
-    values are trustworthy, and exactly the "placeholder read as a confirmed
-    fact" bug ``Fact[T]`` exists to prevent when it doesn't. This is the one
-    correction for that whole class: :func:`apply_legacy_fact_backfill` (the
-    three fields ADR-063 Phase 0 converted) is a thin wrapper over it, and
-    every case-(a) field a later batch converts adds a rule rather than
-    another hand-written loop.
-
-    Only ever *downgrades*, and only for a document that predates the
-    field's own conversion: a v(N)+ document's ``<field>_fact`` was decoded
-    explicitly at construction time and is authoritative.
-    """
-    for rule in rules:
-        if schema_version >= rule.min_schema_version or rule.reliable:
-            continue
-        fact_key = f"{rule.field}_fact"
-        for raw, obj in _owner_pairs(d, rule.owner, decoded):
-            if fact_key in raw:
-                continue
-            setattr(obj, rule.field, rule.normalized_default)
-            setattr(obj, fact_key, Fact.not_collected())
-
-
-def apply_legacy_fact_backfill(
-    d: dict[str, Any],
-    types: list[RecordType],
-    funcs: list[Function],
-    schema_version: int,
-    clang_vtable_facts_reliable_value: bool,
-    clang_va_list_facts_reliable_value: bool,
-    ast_producer_value: str | None,
-    *,
-    variables: list[Any] | None = None,
-    enums: list[Any] | None = None,
-    header_cv_facts_reliable_value: bool = True,
-    clang_restrict_facts_reliable_value: bool = True,
-    castxml_var_access_facts_reliable_value: bool = True,
-    clang_field_initializer_facts_reliable_value: bool = True,
-    clang_deprecation_facts_reliable_value: bool = True,
-) -> None:
-    """Correct the legacy backfill for every case-(a) fact a document predates.
-
-    A pre-v26 snapshot carries no ``vtable_fact``/``vptr_offset_bits_fact``/
-    ``is_va_list_fact`` keys at all, so each ``RecordType``/``Param``'s own
-    ``__post_init__`` bridge already backfilled these to
-    ``Fact.present(raw_value)`` unconditionally (there is no sentinel to
-    distinguish "legacy, key absent" from "legacy, key present" here — both
-    look like an ordinary explicit value to that bridge). That is correct
-    for ``bases``/``virtual_bases`` (no independent reliability signal —
-    see AGENTS.md's ``type_base_changed`` entry), but wrong for
-    ``vtable``/``vptr_offset_bits``/``is_va_list`` when the *existing*
-    reliability flags say this producer's own facts for this snapshot are
-    untrustworthy: ``Fact.present(raw)`` would misread a placeholder value
-    as a confirmed fact, exactly the bug this phase exists to make
-    unrepresentable. Only runs for a legacy (pre-v26) load — a fresh v26+
-    snapshot's ``*_fact`` keys were decoded explicitly at construction time
-    via :func:`decode_fact` and must not be overridden here.
-
-    Phase 5's own case-(a) batches extend the same correction to the fields
-    they convert, each with its own ``min_schema_version`` and its own
-    guarding flag (``header_cv_facts_reliable_value`` for ``TypeField``'s
-    CV facts, schema v38) — one rule added to the tuple below, never a
-    second hand-written loop. The keyword-only spelling keeps every
-    pre-existing caller (and every test constructing this call) unchanged:
-    a flag left at its default ``True`` states "trustworthy", which is what
-    a caller that never heard of that field was already asserting by not
-    correcting it at all.
-
-    ``is_va_list`` needs an extra gate ``vtable``/``vptr_offset_bits`` don't
-    (Codex review, fresh evidence): CastXML never determines va_list-ness at
-    all — its own ``is_va_list`` is always a blanket ``False`` placeholder,
-    not a computed fact the way CastXML's vtable *is* one (see
-    ``clang_vtable_facts_reliable_value``'s own computation in
-    ``serialization.py``: "a castxml... snapshot's own vtable extraction
-    predates this field entirely, so it's always reliable"). But
-    ``clang_va_list_facts_reliable_value`` reads ``True`` for a CastXML
-    snapshot too, since that flag's actual meaning is "safe to trust
-    `False` as not-wrong" (CastXML never reports a real va_list parameter
-    as anything but `False`, so the polarity is never wrong) — a different
-    question from "was this fact actually collected". Reusing that flag
-    alone would silently turn "never observed" into "confirmed not
-    va_list" on every legacy CastXML load. Gated here on
-    ``ast_producer_value == "clang"`` in addition to the reliability flag,
-    so only an actual clang-family load can reach ``Fact.present(...)``.
-    """
-    apply_case_a_fact_backfill(
-        d,
-        schema_version=schema_version,
-        rules=(
-            CaseAFactRule(
-                "RecordType",
-                "vtable",
-                _FACT_FIELDS_SCHEMA_VERSION,
-                clang_vtable_facts_reliable_value,
-                [],
-            ),
-            CaseAFactRule(
-                "RecordType",
-                "vptr_offset_bits",
-                _FACT_FIELDS_SCHEMA_VERSION,
-                clang_vtable_facts_reliable_value,
-                None,
-            ),
-            CaseAFactRule(
-                "Param",
-                "is_va_list",
-                _FACT_FIELDS_SCHEMA_VERSION,
-                ast_producer_value == "clang" and clang_va_list_facts_reliable_value,
-                False,
-            ),
-            # ADR-063 Phase 5 (eighth batch, schema v38): TypeField's own CV
-            # facts. A pre-v38 document carries no is_const_fact/
-            # is_volatile_fact/is_mutable_fact key, so its blanket False
-            # values were bridged to Fact.present(False);
-            # header_cv_facts_reliable is exactly the signal saying whether
-            # that reading is a real fact or a pre-fix castxml placeholder.
-            CaseAFactRule(
-                "TypeField",
-                "is_const",
-                _MIN_SCHEMA_VERSION_FOR_TYPEFIELD_CV_FACTS,
-                header_cv_facts_reliable_value,
-                False,
-            ),
-            CaseAFactRule(
-                "TypeField",
-                "is_volatile",
-                _MIN_SCHEMA_VERSION_FOR_TYPEFIELD_CV_FACTS,
-                header_cv_facts_reliable_value,
-                False,
-            ),
-            CaseAFactRule(
-                "TypeField",
-                "is_mutable",
-                _MIN_SCHEMA_VERSION_FOR_TYPEFIELD_CV_FACTS,
-                header_cv_facts_reliable_value,
-                False,
-            ),
-            # TypeField's other two case-(a) fields, each with its own
-            # guarding flag: a pre-v38 clang document's blanket `None`
-            # default-initializer/deprecation is the same placeholder shape
-            # the CV facts have, and the same two flags the detectors
-            # already consult say so.
-            CaseAFactRule(
-                "TypeField",
-                "default",
-                _MIN_SCHEMA_VERSION_FOR_TYPEFIELD_VALUE_FACTS,
-                clang_field_initializer_facts_reliable_value,
-                None,
-            ),
-            CaseAFactRule(
-                "TypeField",
-                "deprecated",
-                _MIN_SCHEMA_VERSION_FOR_TYPEFIELD_VALUE_FACTS,
-                clang_deprecation_facts_reliable_value,
-                None,
-            ),
-            # ADR-063 Phase 5 (ninth batch, schema v39): the other four
-            # `deprecated` surfaces plus EnumType.is_scoped, all guarded by
-            # the same flag TypeField.deprecated is -- one rule each, which
-            # is the whole point of the rule table.
-            CaseAFactRule(
-                "Function",
-                "deprecated",
-                _MIN_SCHEMA_VERSION_FOR_DEPRECATION_FACTS,
-                clang_deprecation_facts_reliable_value,
-                None,
-            ),
-            CaseAFactRule(
-                "Variable",
-                "deprecated",
-                _MIN_SCHEMA_VERSION_FOR_DEPRECATION_FACTS,
-                clang_deprecation_facts_reliable_value,
-                None,
-            ),
-            CaseAFactRule(
-                "RecordType",
-                "deprecated",
-                _MIN_SCHEMA_VERSION_FOR_DEPRECATION_FACTS,
-                clang_deprecation_facts_reliable_value,
-                None,
-            ),
-            CaseAFactRule(
-                "EnumType",
-                "deprecated",
-                _MIN_SCHEMA_VERSION_FOR_DEPRECATION_FACTS,
-                clang_deprecation_facts_reliable_value,
-                None,
-            ),
-            CaseAFactRule(
-                "EnumType",
-                "is_scoped",
-                _MIN_SCHEMA_VERSION_FOR_DEPRECATION_FACTS,
-                clang_deprecation_facts_reliable_value,
-                None,
-            ),
-            # ADR-063 Phase 5 (tenth batch, schema v40): the last two
-            # case-(a) fields, each with its own guarding flag.
-            CaseAFactRule(
-                "Param",
-                "is_restrict",
-                _MIN_SCHEMA_VERSION_FOR_LAST_CASE_A_FACTS,
-                clang_restrict_facts_reliable_value,
-                False,
-            ),
-            CaseAFactRule(
-                "Variable",
-                "access",
-                _MIN_SCHEMA_VERSION_FOR_LAST_CASE_A_FACTS,
-                castxml_var_access_facts_reliable_value,
-                AccessLevel.PUBLIC,
-            ),
-        ),
-        types=types,
-        functions=funcs,
-        variables=variables or [],
-        enums=enums or [],
-    )
